# perf: optimize AST traversal and add parallel processing