--- a/pyproject.toml
+++ b/pyproject.toml
@@ -29,6 +29,7 @@ dev = [
     "ruff",
     "pytest>=7.3.1",
     "pytest-asyncio>=0.18.2",
+    "pyinstrument>=5.0.3",
 ]
 
 [project.scripts]